+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Create: Mobile Packages - v0.7.7 - 1.21.1 - unreleased
 - Fix Playernetworks Screen now showing network connection status: Member / Owner (#346)
 - Fix NetworkSettingsScreen scrollbar not working (#328)
 - Fix Lock / Unlock button not updating (#331)
+- PlayerNetworksScreen now shows join/leave button depending on membership
 
 ------------------------------------------------------
 Create: Mobile Packages - v0.7.6 - 1.21.1 - 08.05.2026